backend/requirements.txt
flask==1.1.1
flask_restplus==0.13
python-box==5.4.1
PyYAML==5.4.1
Werkzeug==0.16.1  # https://github.com/noirbizarre/flask-restplus/issues/777
